chore: split Type from TypeData
