+++ crates/biome_js_type_info/src/format_type_info.rs
@@ -630,7 +630,6 @@ impl Format<FormatTypeContext> for TypeReference {
                 }
             }
             Self::Import(import) => write!(f, [import.as_ref()]),
-            Self::Unknown => write!(f, [text("unknown reference")]),
         }
     }
 }
