+++ crates/biome_js_type_info/src/type_data/literal.rs
